debt(GraphQL): Remove + deprecate
