+++ server/graphql/v2/mutation/TransactionMutations.ts
@@ -8,9 +8,8 @@ import { purgeCacheForCollective } from '../../../lib/cache';
 import twoFactorAuthLib from '../../../lib/two-factor-authentication';
 import models from '../../../models';
 import { checkRemoteUserCanUseTransactions } from '../../common/scope-check';
-import { canReject } from '../../common/transactions';
+import { canReject, refundTransaction } from '../../common/transactions';
 import { Forbidden, NotFound, Unauthorized, ValidationFailed } from '../../errors';
-import { refundTransaction as legacyRefundTransaction } from '../../v1/mutations/orders';
 import { AmountInput, getValueInCentsFromAmountInput } from '../input/AmountInput';
 import { fetchTransactionWithReference, TransactionReferenceInput } from '../input/TransactionReferenceInput';
 import { Transaction } from '../interface/Transaction';
@@ -86,8 +85,8 @@ const transactionMutations = {
     },
     async resolve(_: void, args, req: express.Request): Promise<typeof Transaction> {
       checkRemoteUserCanUseTransactions(req);
-      const transaction = await fetchTransactionWithReference(args.transaction);
-      return legacyRefundTransaction(undefined, { id: transaction.id }, req);
+      const transaction = await fetchTransactionWithReference(args.transaction, { throwIfMissing: true });
+      return refundTransaction(transaction, req);
     },
   },
   rejectTransaction: {
@@ -127,7 +126,7 @@ const transactionMutations = {
       let refundedTransaction;
       if (!transaction.RefundTransactionId) {
         const refundParams = { id: transaction.id, message: rejectionReason };
-        refundedTransaction = await legacyRefundTransaction(undefined, refundParams, req);
+        refundedTransaction = await refundTransaction(transaction, req, refundParams);
       } else {
         refundedTransaction = await fetchTransactionWithReference({ legacyId: transaction.RefundTransactionId });
       }
