+++ server/graphql/v2/object/Conversation.js
@@ -24,24 +24,10 @@ const Conversation = new GraphQLObjectType({
       updatedAt: { type: new GraphQLNonNull(GraphQLDateTime) },
       tags: { type: new GraphQLList(GraphQLString) },
       summary: { type: new GraphQLNonNull(GraphQLString) },
-      collective: {
-        type: Account,
-        deprecationReason: '2022-09-14: Please use account',
-        resolve(conversation, args, req) {
-          return req.loaders.Collective.byId.load(conversation.CollectiveId);
-        },
-      },
       account: {
         type: Account,
         resolve(conversation, args, req) {
           return req.loaders.Collective.byId.load(conversation.CollectiveId);
-        },
-      },
-      fromCollective: {
-        type: Account,
-        deprecationReason: '2022-09-14: Please use fromAccount',
-        resolve(conversation, args, req) {
-          return req.loaders.Collective.byId.load(conversation.FromCollectiveId);
         },
       },
       fromAccount: {
